Evaluation/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
sqlalchemy>=1.4.0
pymysql>=1.0.0
openai>=1.0.0
python-dotenv>=0.19.0
pyyaml>=6.0
pymilvus>=2.3.0
sqlparse>=0.4.0 